add basePage to reduce code duplication
