+++ src/test/java/page/AbstractPage.java
@@ -4,6 +4,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.pagefactory.AjaxElementLocatorFactory;
 
+
 public abstract class AbstractPage {
     protected WebDriver driver;
