+++ src/test/java/page/CalendarFrame.java
@@ -3,7 +3,6 @@ package page;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
@@ -12,17 +11,12 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CalendarFrame extends AbstractPage {
+public class CalendarFrame extends BasePage {
 
     List<String> timeList = new ArrayList<>();
 
     protected CalendarFrame(WebDriver driver) {
         super(driver);
-    }
-
-    @Override
-    protected AbstractPage openPage() {
-        return this;
     }
 
     @FindBy(css="li.saveSpace")
