+++ src/test/java/page/CalendarPage.java
@@ -5,12 +5,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.FindBy;
 
-public class CalendarPage extends AbstractPage {
-
-    @Override
-    protected AbstractPage openPage() {
-        return this;
-    }
+public class CalendarPage extends BasePage {
 
     public CalendarPage(WebDriver driver) {
         super(driver);
@@ -22,20 +17,16 @@ public class CalendarPage extends AbstractPage {
     @FindBy(xpath = "//a[contains(text(),'here')]")
     WebElement here;
 
-    @FindBy(id = "risk-block")
-    WebElement riskBlock;
-
-
     public CalendarFrame openCalendarFrame() {
         Actions actions = new Actions(driver);
-        actions.scrollToElement(calendarFrame).scrollByAmount(0, riskBlock.getSize().getHeight()).perform();
+        navigate(calendarFrame);
         driver.switchTo().frame(calendarFrame);
         return new CalendarFrame(driver);
     }
 
     public RiskWarningPage clickHere() {
         Actions actions = new Actions(driver);
-        actions.scrollToElement(here).scrollByAmount(0,riskBlock.getSize().getHeight()).perform();
+        navigate(here);
         here.click();
         return new RiskWarningPage(driver);
     }
